+++ commons/ava_services/gecko_service_initializer_core.go
@@ -9,6 +9,7 @@ import (
 	"encoding/pem"
 	"fmt"
 	"github.com/docker/go-connections/nat"
+	"github.com/kurtosis-tech/ava-e2e-tests/commons/ava_default_testnet"
 	"github.com/kurtosis-tech/kurtosis/commons/services"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
@@ -114,9 +115,10 @@ func (g GeckoServiceInitializerCore) InitializeMountedFiles(osFiles map[string]*
 
 	certFilePointer := osFiles[stakingTlsCertFileId]
 	keyFilePointer := osFiles[stakingTlsKeyFileId]
+	defaultStakers := ava_default_testnet.DefaultTestNet.DefaultStakers
 	if len(dependencies) < 5 {
-		certFilePointer.WriteString(DefaultStakers[len(dependencies)].TlsCert)
-		keyFilePointer.WriteString(DefaultStakers[len(dependencies)].PrivateKey)
+		certFilePointer.WriteString(defaultStakers[len(dependencies)].TlsCert)
+		keyFilePointer.WriteString(defaultStakers[len(dependencies)].PrivateKey)
 	} else {
 		rootCA := getRootCA()
 		serviceCert := getServiceCert()
@@ -166,12 +168,13 @@ func (g GeckoServiceInitializerCore) GetStartCommand(mountedFileFilepaths map[st
 			avaDependencies = append(avaDependencies, service.(AvaService))
 		}
 
+		defaultStakers := ava_default_testnet.DefaultTestNet.DefaultStakers
 		socketStrs := make([]string, 0, len(avaDependencies))
 		bootstrapIds := make([]string, 0, len(avaDependencies))
 		for i, service := range avaDependencies {
 			socket := service.GetStakingSocket()
 			socketStrs = append(socketStrs, fmt.Sprintf("%s:%d", socket.GetIpAddr(), socket.GetPort().Int()))
-			bootstrapIds = append(bootstrapIds, DefaultStakers[i].NodeID)
+			bootstrapIds = append(bootstrapIds, defaultStakers[i].NodeID)
 		}
 		if g.stakingTlsEnabled {
 			commandList = append(commandList, "--bootstrap-ids=" + strings.Join(bootstrapIds, ","))