+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Split Gecko networks into MutableGeckoNetwork and FixedGeckoNetwork
 * Fix bug with tests passing when they shouldn't
 * Catch RPC-level errors in MakeRPCRequest
+* Add all five default stakers to staking network bootstrapping
 
 # 0.2.1
 * Fix tee suppressing exit code of the Docker image